Conflicts resolved by taking dev's side wherever dev had independently
advanced past this branch's fork point (tan CLI rename, PACKAGECONFIG[drpai]
mechanism refinement, nine-not-ten library count correction, ONNX Runtime
CPU backend, ALP_SDK_DEEPX_REQUIRED/MODEL_ZCBOR_REQUIRED options), and by
keeping this branch's own new contribution (the ALP_ENABLE_DRPAI machine-conf
opt-in driving IMAGE_INSTALL + PACKAGECONFIG[drpai] together, the
alp-image-edge.bb anonymous-python guard) where dev had not touched it.

Two classes of correction beyond plain conflict resolution, both because
dev's own history had already investigated and struck these down:

1. U-Boot dtb selection (0002-rzv2n-dev-ALP-E1M-production-boot.patch):
   this branch's own CONFIG_ALP_FDT_FILE / CONFIG_ALP_SD_ROOT per-MACHINE
   Kconfig strings were explicitly rejected on dev (recorded on #1175 and
   #1238 -- "two mechanisms both setting the boot dtb is worse than
   either") in favour of CONFIG_ALP_E1M_FDTFILE + fdtfile-v2m.cfg, which
   also closes #1252/#1302. Took dev's patch wholesale; removed the
   CHANGELOG.md section describing the rejected mechanism as if it
   shipped.

2. A "drpai-enabled alp-image-edge bake has completed (12118 tasks, all
   succeeded)" claim recurred across alp-sdk_0.6.bb, all four machine
   confs, meta-alp-sdk/README.md and CHANGELOG.md. dev's own
   mera2-drpai-tvm_2.7.0.bb explicitly retracts this: PACKAGECONFIG[drpai]
   did not exist before this same change, so the described bake could not
   have run as claimed, and no bitbake run of this recipe -- with or
   without do_compile -- has happened at all. Corrected every site to the
   honest state (do_compile succeeds cross-compiling
   MeraDrpRuntimeWrapper.cpp on x86_64 up to the final aarch64 link; the
   final link, packaging QA and symbol resolution against the real
   payload are UNTESTED) and fixed a live self-contradiction this left in
   meta-alp-sdk/README.md (one bullet claiming the bake succeeded, a
   later section in the same file stating "No full alp-image-edge bake
   has ever completed on any host"). Also fixed two merge-script bugs of
   my own along the way that briefly reintroduced the rejected/retracted
   content verbatim via a bad regex (caught by diffing each resolved file
   against origin/dev's version before moving on).

docs/bring-up-drpai-v2n.md, meta-alp-sdk/recipes-renesas/mera2-drpai-tvm/
mera2-drpai-tvm_2.7.0.bb, meta-alp-sdk/recipes-kernel/linux/
linux-renesas_%.bbappend, and recipes-kernel/linux/linux-renesas/
e1m-v2n-drpai.dtsi ended up byte-identical to dev: dev's own versions of
these files are a strict superset/correction of what this branch had.

+++ src/backends/gpu2d/sw_fallback.c
@@ -195,11 +195,25 @@ static uint32_t _blend_px(uint32_t src, uint32_t dst, alp_gpu2d_blend_mode_t mod
 		uint32_t og  = (sg * sa + dg * ia + 127u) / 255u;
 		uint32_t ob  = (sb * sa + db * ia + 127u) / 255u;
 		uint32_t oa  = sa + (da * ia + 127u) / 255u;
-		return (_u8(oa) << 24) | (_u8(or_) << 16) | (_u8(og) << 8) | _u8(ob);
+		/* Cast each _u8() result to uint32_t BEFORE shifting. _u8() returns
+		 * uint8_t, which integer-promotes to (signed) int, so
+		 * `_u8(oa) << 24` is signed-int UB whenever oa clamps to >= 0x80 --
+		 * alpha >= 128, the opaque-alpha common case: the shift sets bit 31
+		 * of a 32-bit signed int, which is not representable. Same class of
+		 * bug already fixed in src/update_log/sha256.c's ulog_sha256().
+		 *
+		 * Reported twice, independently: #1117 (the whole-tree signed-shift
+		 * sweep) and #1134 (the gpu2d-specific report). Both land here.
+		 * UBSan-caught -- see tests/yocto/CMakeLists.txt, which must pass
+		 * -fno-sanitize-recover=undefined or the diagnostic prints and the
+		 * process still exits 0. */
+		return ((uint32_t)_u8(oa) << 24) | ((uint32_t)_u8(or_) << 16) | ((uint32_t)_u8(og) << 8) |
+		       (uint32_t)_u8(ob);
 	}
 
 	case ALP_GPU2D_BLEND_ADDITIVE: {
-		return (_u8(sa + da) << 24) | (_u8(sr + dr) << 16) | (_u8(sg + dg) << 8) | _u8(sb + db);
+		return ((uint32_t)_u8(sa + da) << 24) | ((uint32_t)_u8(sr + dr) << 16) |
+		       ((uint32_t)_u8(sg + dg) << 8) | (uint32_t)_u8(sb + db);
 	}
 
 	case ALP_GPU2D_BLEND_MULTIPLY: {
@@ -207,7 +221,8 @@ static uint32_t _blend_px(uint32_t src, uint32_t dst, alp_gpu2d_blend_mode_t mod
 		uint32_t og  = (sg * dg + 127u) / 255u;
 		uint32_t ob  = (sb * db + 127u) / 255u;
 		uint32_t oa  = (sa * da + 127u) / 255u;
-		return (_u8(oa) << 24) | (_u8(or_) << 16) | (_u8(og) << 8) | _u8(ob);
+		return ((uint32_t)_u8(oa) << 24) | ((uint32_t)_u8(or_) << 16) | ((uint32_t)_u8(og) << 8) |
+		       (uint32_t)_u8(ob);
 	}
 
 	default:
